+++ dbsnp/tests/factories.py
@@ -9,7 +9,7 @@ class DbsnpFactory(factory.django.DjangoModelFactory):
 
     release = "GRCh37"
     chromosome = factory.Iterator(list(map(str, range(1, 23))) + ["X", "Y"])
-    position = factory.Sequence(lambda n: n * 100)
+    position = factory.Sequence(lambda n: (n + 1) * 100)
     reference = factory.Iterator("ACGT")
     alternative = factory.Iterator("CGTA")
     rsid = factory.Sequence(lambda n: "rs%d" % n)